+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JWB1XSRXWZ0R4RD9SF9QXMSH"
+        "01JWB9ZZ2YTFW9AZ0B31TDGKRC"
     )
     r"""The id of the resource"""