+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JWB1XSRWWVWM796XCPWA5T35"
+        "01JWB9ZZ2YY8BDM6GK5CRKGW0X"
     )
     r"""The id of the resource"""