+++ packages/orq-rc/RELEASES.md
@@ -1569,3 +1569,13 @@ Based on:
 - [python v3.6.0-rc.42] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.42] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.42 - packages/orq-rc
+
+## 2025-05-28 11:22:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.552.0 (2.610.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.6.0-rc.43] packages/orq-rc
+### Releases
+- [PyPI v3.6.0-rc.43] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.43 - packages/orq-rc